TSA - move Time Filters below TSA Card
